+++ SampleApp/Sources/Sample/Components/Tag/TagViewController.swift
@@ -7,7 +7,7 @@ class TagViewController: UIViewController, SampleItem {
     private let defaultTagView = NatTag(style: .defaultAlert)
     private let leftTagView = NatTag(style: .leftAlert)
     private let rightTagView = NatTag(style: .rightAlert)
-    
+
     private lazy var stackView: UIStackView = {
         let stackView = UIStackView()
         stackView.axis = .vertical
@@ -34,12 +34,12 @@ class TagViewController: UIViewController, SampleItem {
 
         let constraints = [
             stackView.centerYAnchor.constraint(equalTo: view.centerYAnchor),
-            stackView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            stackView.centerXAnchor.constraint(equalTo: view.centerXAnchor)
         ]
 
         NSLayoutConstraint.activate(constraints)
     }
-    
+
     private func createTag(style: NatTag.Style, text: String) -> NatTag {
         let tag = NatTag(style: style)
         tag.configure(text: text)